Conflicts:
	libtransmission/peer-mgr.c

+++ libtransmission/peer-mgr.c
@@ -804,11 +804,11 @@ comparePieceByWeight( const void * va, const void * vb )
     missing = tr_cpMissingBlocksInPiece( &tor->completion, a->index );
     pending = a->requestCount;
     ia = missing > pending ? missing - pending
-        : (int) ( tr_torPieceCountBlocks( tor, a->index ) + pending );
+        : ( tr_torPieceCountBlocks( tor, a->index ) + pending );
     missing = tr_cpMissingBlocksInPiece( &tor->completion, b->index );
     pending = b->requestCount;
     ib = missing > pending ? missing - pending
-        : (int) ( tr_torPieceCountBlocks( tor, b->index ) + pending );
+        : ( tr_torPieceCountBlocks( tor, b->index ) + pending );
     if( ia < ib ) return -1;
     if( ia > ib ) return 1;
